requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 24.9.2
_openmp_mutex=4.5=2_gnu
biopython=1.85=pypi_0
blas=1.0=mkl
brotli-python=1.0.9=py313h5da7b33_9
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2025.2.25=haa95532_0
contourpy=1.3.1=py313h214f63a_0
cycler=0.11.0=pyhd3eb1b0_0
ete3=3.1.3=pyhd8ed1ab_0
expat=2.6.4=h8ddb27b_0
fonttools=4.55.3=py313h827c3e9_0
freetype=2.12.1=ha860e81_0
icc_rt=2022.1.0=h6049295_2
icu=73.1=h6c2663c_0
intel-openmp=2023.1.0=h59b6b97_46320
jpeg=9e=h827c3e9_3
kiwisolver=1.4.8=py313h5da7b33_0
krb5=1.20.1=h5b6d351_0
lcms2=2.16=hb4a4139_0
legacy-cgi=2.6.3=pypi_0
lerc=4.0.0=h5da7b33_0
libdeflate=1.22=h5bf469e_0
libffi=3.4.4=hd77b12b_1
libgcc=14.2.0=h1383e82_2
libgomp=14.2.0=h1383e82_2
libiconv=1.16=h2bbff1b_3
libmpdec=4.0.0=h827c3e9_0
libpng=1.6.39=h8cc25b3_0
libpq=17.4=h70ee33d_0
libtiff=4.5.1=h44ae7cf_1
libwebp-base=1.3.2=h3d04722_1
libwinpthread=12.0.0.r4.gg4f2fc60ca=h57928b3_9
libxcb=1.17.0=h0e4246c_0
libxml2=2.13.5=h24da03e_0
libxslt=1.1.41=h0739af5_0
lxml=5.3.0=py313h395c83e_1
lz4-c=1.9.4=h2bbff1b_1
matplotlib=3.10.0=py313haa95532_1
matplotlib-base=3.10.0=py313h7aa5d4e_1
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py313h827c3e9_2
mkl_fft=1.3.11=py313h827c3e9_0
mkl_random=1.2.8=py313hce38976_0
numpy=2.2.4=py313he6dc315_0
numpy-base=2.2.4=py313h6011491_0
openjpeg=2.5.2=hae555c5_0
openssl=3.0.16=h3f729d1_0
packaging=24.2=py313haa95532_0
pillow=11.1.0=py313h096bfcc_0
pip=25.0=py313haa95532_0
pthread-stubs=0.3=h3c9f919_1
pyparsing=3.2.0=py313haa95532_0
pyqt=6.7.1=py313h5da7b33_0
pyqt5=5.15.11=pypi_0
pyqt5-qt5=5.15.2=pypi_0
pyqt5-sip=12.17.0=pypi_0
pyqt6-sip=13.9.1=py313h827c3e9_0
python=3.13.2=hadb2040_100_cp313
python-dateutil=2.9.0post0=py313haa95532_2
python_abi=3.13=0_cp313
qtbase=6.7.3=h0804d20_0
qtdeclarative=6.7.3=h5da7b33_0
qtsvg=6.7.3=hf2fb9eb_0
qttools=6.7.2=h0de5f00_0
qtwebchannel=6.7.3=h5da7b33_0
qtwebsockets=6.7.3=h5da7b33_0
scipy=1.15.1=py313hd4070db_0
setuptools=72.1.0=py313haa95532_0
sip=6.10.0=py313h5da7b33_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.45.3=h2bbff1b_0
tbb=2021.8.0=h59b6b97_0
tk=8.6.14=h0416ee5_0
tornado=6.4.2=py313h827c3e9_0
tzdata=2025a=h04d1e81_0
ucrt=10.0.20348.0=haa95532_0
vc=14.42=haa95532_4
vs2015_runtime=14.42.34433=he0abc0d_4
wheel=0.45.1=py313haa95532_0
xorg-libice=1.1.2=h0e40799_0
xorg-libsm=1.2.6=h0e40799_0
xorg-libx11=1.8.12=hf48077a_0
xorg-libxau=1.0.12=h0e40799_0
xorg-libxdmcp=1.1.5=h0e40799_0
xorg-libxext=1.3.6=h0e40799_0
xorg-libxrender=0.9.12=h0e40799_0
xorg-xextproto=7.3.0=h0e40799_1004
xz=5.6.4=h4754444_1
zlib=1.2.13=h8cc25b3_1
zstd=1.5.6=h8880b57_0
